rename web_browser to _web_browser
